requirements.txt
pytest
pytest-html
requests
pyyaml